New lighting files
Changed barrel files to a single file passing the true/false variable
from fn_selfActions.
Added Tower Lighting toggle.
Both now look for a flag within 200 meters of the panel and then look
for barrels/towers within 200 meters of the flag. This will prevent some
barrels/towers not toggling if they are not within 200 meters of the
panel itself.