New Reposition System
Players now select "Preview" instead of finish. At that stage they can
either confirm the build and go ahead as normal, or reposition the
build.

If repositioning, it will exit player_build, delete the object but store
it's position and direction, then relaunch player_build with the
relevant information. Should make those annoying builds that move a
little easier to deal with.